externalize plugin
