+++ scripts/generate-md-exports.mjs
@@ -26,7 +26,8 @@ import RemarkLinkRewrite from 'remark-link-rewrite';
 import remarkStringify from 'remark-stringify';
 import {unified} from 'unified';
 import {remove} from 'unist-util-remove';
-import {visit} from 'unist-util-visit';
+
+import {rehypeExpandCodeTabs} from './rehype-expand-code-tabs.mjs';
 
 const DOCS_ORIGIN = process.env.NEXT_PUBLIC_DEVELOPER_DOCS
   ? 'https://develop.sentry.dev'
@@ -956,71 +957,6 @@ function extractContentForCacheKey(html) {
     .replace(/(\w+__)[a-zA-Z0-9]{5}/g, '$1X');
 
   return title + '\0' + canonical + '\0' + normalizedMain;
-}
-
-/**
- * Rehype plugin that expands CodeTabs panels for markdown export.
- *
- * CodeTabs renders all tab panels in the DOM but hides non-active ones with
- * the `hidden` attribute. Each panel carries `data-code-tab-title` and
- * optionally `data-code-tab-filename`. This plugin:
- *  1. Removes the `hidden` attribute so rehypeRemark processes every panel
- *  2. Strips CodeBlock chrome (filename display, copy button) from the panel
- *  3. Inserts a bold heading (preferring filename, falling back to tab title)
- *     before the <pre> block so readers can tell which file the snippet
- *     belongs to
- */
-export function rehypeExpandCodeTabs() {
-  return tree => {
-    visit(tree, 'element', node => {
-      const title = node.properties?.dataCodeTabTitle;
-      if (!title) {
-        return;
-      }
-      delete node.properties.hidden;
-      const filename = node.properties?.dataCodeTabFilename;
-      delete node.properties.dataCodeTabTitle;
-      delete node.properties.dataCodeTabFilename;
-
-      const label = filename || title;
-      const preElements = collectAll(node, el => el.tagName === 'pre');
-      if (preElements.length === 0) {
-        return;
-      }
-
-      const heading = {
-        type: 'element',
-        tagName: 'p',
-        properties: {},
-        children: [
-          {
-            type: 'element',
-            tagName: 'strong',
-            properties: {},
-            children: [{type: 'text', value: label}],
-          },
-        ],
-      };
-
-      node.children = [heading, ...preElements];
-    });
-  };
-}
-
-function collectAll(node, predicate) {
-  const results = [];
-  (function walk(n) {
-    if (n.type === 'element' && predicate(n)) {
-      results.push(n);
-      return;
-    }
-    if (n.children) {
-      for (const child of n.children) {
-        walk(child);
-      }
-    }
-  })(node);
-  return results;
 }
 
 async function genMDFromHTML(source, {cacheDir, noCache, usedCacheFiles}) {
